TGP-3177: update goodsRecord controller - split into 2 separate update and create controllers and move into files
