+++ app/controllers/goodsRecord/productReference/CreateProductReferenceController.scala
@@ -14,27 +14,25 @@
  * limitations under the License.
  */
 
-package controllers.goodsRecord
+package controllers.goodsRecord.productReference
 
 import connectors.GoodsRecordConnector
 import controllers.BaseController
 import controllers.actions._
 import forms.goodsRecord.ProductReferenceFormProvider
 import models.Mode
-import models.helper.GoodsDetailsUpdate
 import navigation.GoodsRecordNavigator
-import pages.goodsRecord.{ProductReferencePage, ProductReferenceUpdatePage}
+import pages.goodsRecord.ProductReferencePage
 import play.api.i18n.MessagesApi
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import services.AuditService
-import utils.SessionData._
 import views.html.goodsRecord.ProductReferenceView
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class ProductReferenceController @Inject() (
+class CreateProductReferenceController @Inject() (
   override val messagesApi: MessagesApi,
   sessionRepository: SessionRepository,
   navigator: GoodsRecordNavigator,
@@ -50,35 +48,22 @@ class ProductReferenceController @Inject() (
 )(implicit ec: ExecutionContext)
     extends BaseController {
 
-  private val form                                     = formProvider()
-  def onPageLoadCreate(mode: Mode): Action[AnyContent] =
+  private val form = formProvider()
+
+  def onPageLoad(mode: Mode): Action[AnyContent] =
     (identify andThen profileAuth andThen getData andThen requireData) { implicit request =>
       val preparedForm = prepareForm(ProductReferencePage, form)
 
-      val onSubmitAction = controllers.goodsRecord.routes.ProductReferenceController.onSubmitCreate(mode)
+      val onSubmitAction =
+        controllers.goodsRecord.productReference.routes.CreateProductReferenceController.onSubmit(mode)
 
       Ok(view(preparedForm, onSubmitAction))
     }
 
-  def onPageLoadUpdate(mode: Mode, recordId: String): Action[AnyContent] =
-    (identify andThen profileAuth andThen getData andThen requireData) { implicit request =>
-      val preparedForm = prepareForm(ProductReferenceUpdatePage(recordId), form)
-
-      auditService
-        .auditStartUpdateGoodsRecord(
-          request.eori,
-          request.affinityGroup,
-          GoodsDetailsUpdate,
-          recordId
-        )
-
-      val onSubmitAction = controllers.goodsRecord.routes.ProductReferenceController.onSubmitUpdate(mode, recordId)
-      Ok(view(preparedForm, onSubmitAction)).removingFromSession(dataRemoved, dataUpdated, pageUpdated)
-    }
-
-  def onSubmitCreate(mode: Mode): Action[AnyContent] =
+  def onSubmit(mode: Mode): Action[AnyContent] =
     (identify andThen profileAuth andThen getData andThen requireData).async { implicit request =>
-      val onSubmitAction = controllers.goodsRecord.routes.ProductReferenceController.onSubmitCreate(mode)
+      val onSubmitAction =
+        controllers.goodsRecord.productReference.routes.CreateProductReferenceController.onSubmit(mode)
 
       form
         .bindFromRequest()
@@ -99,30 +84,4 @@ class ProductReferenceController @Inject() (
         )
     }
 
-  def onSubmitUpdate(mode: Mode, recordId: String): Action[AnyContent] =
-    (identify andThen profileAuth andThen getData andThen requireData).async { implicit request =>
-      val onSubmitAction = controllers.goodsRecord.routes.ProductReferenceController.onSubmitUpdate(mode, recordId)
-      form
-        .bindFromRequest()
-        .fold(
-          formWithErrors => Future.successful(BadRequest(view(formWithErrors, onSubmitAction))),
-          value =>
-            for {
-              isproductReferenceUnique <- goodsRecordConnector.isproductReferenceUnique(value)
-              oldRecord                <- goodsRecordConnector.getRecord(recordId)
-              updatedAnswers           <-
-                Future.fromTry(request.userAnswers.set(ProductReferenceUpdatePage(recordId), value))
-            } yield
-              if (isproductReferenceUnique || oldRecord.traderRef == value) {
-                sessionRepository.set(updatedAnswers)
-                Redirect(navigator.nextPage(ProductReferenceUpdatePage(recordId), mode, updatedAnswers))
-                  .addingToSession(dataUpdated -> (oldRecord.traderRef != value).toString)
-                  .addingToSession(pageUpdated -> productReference)
-              } else {
-                val formWithApiErrors =
-                  createFormWithErrors(form, value, "productReference.error.traderRefNotUnique")
-                BadRequest(view(formWithApiErrors, onSubmitAction))
-              }
-        )
-    }
 }
