+++ app/navigation/GoodsRecordNavigator.scala
@@ -31,7 +31,7 @@ class GoodsRecordNavigator @Inject() extends Navigator {
 
   override val normalRoutes: Page => UserAnswers => Call = {
     case CreateRecordStartPage            =>
-      _ => controllers.goodsRecord.routes.ProductReferenceController.onPageLoadCreate(NormalMode)
+      _ => controllers.goodsRecord.productReference.routes.CreateProductReferenceController.onPageLoad(NormalMode)
     case ProductReferencePage             =>
       _ => controllers.goodsRecord.routes.GoodsDescriptionController.onPageLoadCreate(NormalMode)
     case p: ProductReferenceUpdatePage    =>
