+++ conf/app.goodsRecord.routes
@@ -37,15 +37,15 @@ POST        /update-record/:recordId/goods-description/check                    
 GET         /create-record/check-your-answers                                                       controllers.goodsRecord.CyaCreateRecordController.onPageLoad()
 POST        /create-record/check-your-answers                                                       controllers.goodsRecord.CyaCreateRecordController.onSubmit()
 
-GET         /create-record/product-reference                                                        controllers.goodsRecord.ProductReferenceController.onPageLoadCreate(mode: Mode = NormalMode)
-POST        /create-record/product-reference                                                        controllers.goodsRecord.ProductReferenceController.onSubmitCreate(mode: Mode = NormalMode)
-GET         /create-record/product-reference/check                                                  controllers.goodsRecord.ProductReferenceController.onPageLoadCreate(mode: Mode = CheckMode)
-POST        /create-record/product-reference/check                                                  controllers.goodsRecord.ProductReferenceController.onSubmitCreate(mode: Mode = CheckMode)
+GET         /create-record/product-reference                                                        controllers.goodsRecord.productReference.CreateProductReferenceController.onPageLoad(mode: Mode = NormalMode)
+POST        /create-record/product-reference                                                        controllers.goodsRecord.productReference.CreateProductReferenceController.onSubmit(mode: Mode = NormalMode)
+GET         /create-record/product-reference/check                                                  controllers.goodsRecord.productReference.CreateProductReferenceController.onPageLoad(mode: Mode = CheckMode)
+POST        /create-record/product-reference/check                                                  controllers.goodsRecord.productReference.CreateProductReferenceController.onSubmit(mode: Mode = CheckMode)
 
-GET         /update-record/:recordId/product-reference                                              controllers.goodsRecord.ProductReferenceController.onPageLoadUpdate(mode: Mode = NormalMode, recordId: String)
-POST        /update-record/:recordId/product-reference                                              controllers.goodsRecord.ProductReferenceController.onSubmitUpdate(mode: Mode = NormalMode, recordId: String)
-GET         /update-record/:recordId/product-reference/check                                        controllers.goodsRecord.ProductReferenceController.onPageLoadUpdate(mode: Mode = CheckMode, recordId: String)
-POST        /update-record/:recordId/product-reference/check                                        controllers.goodsRecord.ProductReferenceController.onSubmitUpdate(mode: Mode = CheckMode, recordId: String)
+GET         /update-record/:recordId/product-reference                                              controllers.goodsRecord.productReference.UpdateProductReferenceController.onPageLoad(mode: Mode = NormalMode, recordId: String)
+POST        /update-record/:recordId/product-reference                                              controllers.goodsRecord.productReference.UpdateProductReferenceController.onSubmit(mode: Mode = NormalMode, recordId: String)
+GET         /update-record/:recordId/product-reference/check                                        controllers.goodsRecord.productReference.UpdateProductReferenceController.onPageLoad(mode: Mode = CheckMode, recordId: String)
+POST        /update-record/:recordId/product-reference/check                                        controllers.goodsRecord.productReference.UpdateProductReferenceController.onSubmit(mode: Mode = CheckMode, recordId: String)
 
 GET         /goods-record/:recordId                                                                 controllers.goodsRecord.SingleRecordController.onPageLoad(recordId: String)
 
@@ -68,6 +68,7 @@ GET         /update-record/:recordId/check-your-answers/country-of-origin       
 POST        /update-record/:recordId/check-your-answers/country-of-origin                           controllers.goodsRecord.CyaUpdateRecordController.onSubmitCountryOfOrigin(recordId: String)
 GET         /update-record/:recordId/check-your-answers/goods-description                           controllers.goodsRecord.CyaUpdateRecordController.onPageLoadGoodsDescription(recordId: String)
 POST        /update-record/:recordId/check-your-answers/goods-description                           controllers.goodsRecord.CyaUpdateRecordController.onSubmitGoodsDescription(recordId: String)
+
 GET         /update-record/:recordId/check-your-answers/product-reference                           controllers.goodsRecord.CyaUpdateRecordController.onPageLoadproductReference(recordId: String)
 POST        /update-record/:recordId/check-your-answers/product-reference                           controllers.goodsRecord.CyaUpdateRecordController.onSubmitproductReference(recordId: String)
 GET         /update-record/:recordId/check-your-answers/commodity-code                              controllers.goodsRecord.CyaUpdateRecordController.onPageLoadCommodityCode(recordId: String)
